Split reaction length setting in two: maximum display length and maximum storage length.  Reactions longer than the storage length are dropped, rather than truncated, in order to avoid creating a new, distinct reaction when interacting with a reaction from a different client (which may allow longer reactions).
