+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Added:
 
-- Setting to limit reaction text length (`buffer.channel.message.max_reaction_chars`)
+- Setting to limit reaction display length (`buffer.channel.message.max_reaction_display`) and maximum length (`buffer.channel.message.max_reaction_chars`)
 
 Fixed:
