+++ book/src/configuration/buffer/channel/message.md
@@ -6,6 +6,7 @@ Message settings within a channel buffer.
   - [Configuration](#configuration)
     - [nickname\_color](#nickname_color)
     - [show\_emoji\_reacts](#show_emoji_reacts)
+    - [max\_reaction\_display](#max_reaction_display)
     - [max\_reaction\_chars](#max_reaction_chars)
 
 ## Configuration
@@ -36,10 +37,10 @@ Whether to display emoji reactions on messages (if [IRCv3 React](https://ircv3.n
 show_emoji_reacts = true
 ```
 
-### max_reaction_chars
+### max_reaction_display
 
 Maximum number of user-visible characters (Unicode grapheme clusters) in a reaction.
-If a reaction exceeds this value, it is truncated to the first `max_reaction_chars` grapheme clusters.
+If a reaction exceeds this value, then its display is truncated to the first `max_reaction_display` grapheme clusters.
 
 ```toml
 # Type: integer
@@ -47,5 +48,19 @@ If a reaction exceeds this value, it is truncated to the first `max_reaction_cha
 # Default: 5
 
 [buffer.channel.message]
-max_reaction_chars = 5
+max_reaction_display = 5
+```
+
+### max_reaction_chars
+
+Maximum number of user-visible characters (Unicode grapheme clusters) in a reaction.
+If a reaction exceeds this value, then it is not stored.
+
+```toml
+# Type: integer
+# Values: positive integers
+# Default: 64
+
+[buffer.channel.message]
+max_reaction_chars = 64
 ```
