+++ src/buffer/scroll_view.rs
@@ -1164,26 +1164,10 @@ impl State {
                 }
             }
             Message::Reacted { msgid, text } => {
-                send_reaction(
-                    clients,
-                    buffer,
-                    history,
-                    msgid,
-                    text,
-                    false,
-                    config.buffer.channel.message.max_reaction_chars,
-                );
+                send_reaction(clients, buffer, history, msgid, text, false);
             }
             Message::Unreacted { msgid, text } => {
-                send_reaction(
-                    clients,
-                    buffer,
-                    history,
-                    msgid,
-                    text,
-                    true,
-                    config.buffer.channel.message.max_reaction_chars,
-                );
+                send_reaction(clients, buffer, history, msgid, text, true);
             }
         }
         (Task::none(), None)
@@ -1410,10 +1394,8 @@ fn send_reaction(
     msgid: message::Id,
     text: String,
     unreact: bool,
-    max_reaction_chars: u32,
 ) -> Option<()> {
     let buffer = buffer?;
-    let text = reaction::truncate_text(&text, max_reaction_chars as usize);
     let server = buffer.server();
     let target = buffer.target()?;
     let command = match unreact {
